Adiciona busca a ludo.py, modifica test.py
