+++ genetic_algorithm.py
@@ -30,13 +30,15 @@ def evaluate_player(player, num_games=100):
     return num
 
 class GeneticAlgorithm:
-    def __init__(self, population_size, generations, mutation_rate, think_type, num_games, simulation_type):
+    def __init__(self, population_size, generations, mutation_rate, think_type, num_games, simulation_type, crossover_type=1, selection_type=1):
         self.population_size = population_size + (population_size % 4)
         self.generations = generations
         self.mutation_rate = mutation_rate
         self.think_type = think_type
         self.num_games = num_games
         self.simulation_type = simulation_type
+        self.crossover_type = crossover_type
+        self.selection_type = selection_type
 
     def initialize_population(self):
         self.population = [self.random_individual() for _ in range(self.population_size)]
@@ -44,8 +46,12 @@ class GeneticAlgorithm:
     
     def random_individual(self):
 
-        dna = np.random.randn(6)
 
+        if (self.think_type == "arbitrary"):
+            dna = np.random.randn(8)
+        elif (self.think_type == "neural"):
+            dna = (np.random.randn(233, 4), np.random.randn(4))
+            
         random_player = Player(
             name = "",
             think_type = self.think_type
@@ -68,28 +74,71 @@ class GeneticAlgorithm:
         # print([p.wins for p in self.population])
 
         self.population.sort(key=lambda p: p.wins, reverse=True)
-        elite = self.population[:num_parents // 4]
-    
-        # Roleta para o restante
-        probabilities = [p.fitness for p in self.population]
-        selected = list(np.random.choice(
-            self.population, 
-            size=(num_parents // 4)*3, 
-            replace=False, 
-            p=np.array(probabilities) / np.sum(probabilities)
-        ))
+        if (self.selection_type == 1):
+
+            elite = self.population[:num_parents // 4]
         
+            # Roleta para o restante
+            probabilities = [p.fitness for p in self.population]
+            selected = list(np.random.choice(
+                self.population, 
+                size=(num_parents // 4)*3, 
+                replace=True, 
+                p=np.array(probabilities) / np.sum(probabilities)
+            ))
+        elif (self.selection_type == 2):
+        
+            elite = []
+
+            probabilities = [p.fitness for p in self.population]
+            selected = list(np.random.choice(
+                self.population, 
+                size=num_parents, 
+                replace=True, 
+                p=np.array(probabilities) / np.sum(probabilities)
+            ))
+
         return elite + selected
 
     def crossover(self, parent1, parent2):
-        mask = np.random.randint(0, 2, size=parent1.DNA.shape)
-        child_dna = np.where(mask, parent1.DNA, parent2.DNA)
+        if (self.crossover_type == 1):
+            if (self.think_type == "arbitrary"):
+                child_dna = (parent1.DNA + parent2.DNA) / 2
+            elif (self.think_type == "neural"):
+                child_dna = (
+                    (parent1.DNA[0] + parent2.DNA[0]) / 2,
+                    (parent1.DNA[1] + parent2.DNA[1]) / 2
+                )
+        else:
+            if (self.think_type == "arbitrary"):
+                # Crossover por máscara binária (como já estava)
+                mask = np.random.randint(0, 2, size=parent1.DNA.shape)
+                child_dna = np.where(mask, parent1.DNA, parent2.DNA)
+            elif (self.think_type == "neural"):
+                # Crossover por máscara binária para redes neurais
+                mask1 = np.random.randint(0, 2, size=parent1.DNA[0].shape)
+                mask2 = np.random.randint(0, 2, size=parent1.DNA[1].shape)
+                child_dna = (
+                    np.where(mask1, parent1.DNA[0], parent2.DNA[0]),
+                    np.where(mask2, parent1.DNA[1], parent2.DNA[1])
+                )
+
         return child_dna
 
     def mutate(self, dna):
-        for i in range(len(dna)):
-            if random.random() < self.mutation_rate:
-                dna[i] += np.random.normal(0, 0.5)
+
+        if (self.think_type == "arbitrary"):
+            for i in range(len(dna)):
+                if random.random() < self.mutation_rate:
+                    dna[i] += np.random.normal(0, 0.05)
+        elif (self.think_type == "neural"):
+            for i in range(len(dna[0])):
+                if random.random() < self.mutation_rate:
+                    dna[0][i] += np.random.normal(0, 0.05)
+            for i in range(len(dna[1])):
+                if random.random() < self.mutation_rate:
+                    dna[1][i] += np.random.normal(0, 0.05)
+
         return dna
     
     def evolve(self):
@@ -103,10 +152,11 @@ class GeneticAlgorithm:
             self.evaluate_fitness()
             
             self.population.sort(key=lambda p: p.wins, reverse=True)
-            # for player in self.population[:4]:
-            #     print(f"Player DNA: {player.DNA}, Wins: {player.wins}")
+            for player in self.population:
+                print(f"{player.wins}", end=" ")
             
-            print("Player evaluation:", evaluate_player(self.population[0], 100))
+            print("Player evaluation:", (evaluate_player(self.population[0], 1000) + evaluate_player(self.population[1], 1000) + evaluate_player(self.population[2], 1000))/3)
+            # print(self.population[0].DNA)
 
             # Seleciona pais
             parents = self.select_parents(self.population_size)
@@ -139,6 +189,7 @@ class GeneticAlgorithm:
             for i in range(len(self.population) // 4):
 
                 for k in range(self.num_games):
+
                     players = self.population[i * 4:(i + 1) * 4]
                     for j in range(4):
                         players[j].set_oponents(players[:j] + players[j + 1:])
@@ -174,14 +225,16 @@ class GeneticAlgorithm:
 
 
 
-start_time = time.time()
-gen = GeneticAlgorithm(200, 1000, 0.2, "arbitrary", 100, 2)
+# start_time = time.time()
+# gen = GeneticAlgorithm(200, 1000, 0.05, "arbitrary", 100, 2, 1)
+# gen.initialize_population()
+# gen.evolve()
+# end_time = time.time()
+# print(f"Tempo de execução do evolve: {end_time - start_time:.2f} segundos")
+
+
+# print("Player evaluation:", evaluate_player(gen.population[0], 1000))
+
+gen = GeneticAlgorithm(100, 1000, 0.05, "neural", 100, 2, 1, 2)
 gen.initialize_population()
 gen.evolve()
-end_time = time.time()
-print(f"Tempo de execução do evolve: {end_time - start_time:.2f} segundos")
-
-
-print("Player evaluation:", evaluate_player(gen.population[0], 1000))
-
-
